Bug 1381840 - Schedule 1-Day Firefox Retention
This schedules the job using a slack of 4 days and 13 bits for the HLL
object. This corresponds to the observation of 99% of client activity
with roughly a 1% standard error introduced HLL.

This job shares components in mozetl and in telemetry-batch-view, so
there is a custom job script to launch. This script also accepts
arguments from the command-line can be useful for backfill.