Bug 1517013 - Separate telemetry aggregates
Previously, telemetry aggregates were part of the same DAG. However,
they are not dependent on one another; so when one fails, the other
eventually stops running (due to an airflow).

This is not the behavior we want -  we want the other to run when
one is failing. This way they are totally independent on the other
one running.