Use only main ping view in update orphaning ETL
This removes usage of `main_1pct_backfill` table which is no longer
needed after main ping table has been backfilled.

This also adds a filter for `build.version` being not null. Such pings
should not be included in this analysis (we had a single ping having
this field empty in the last year).