Don't delete from GCS in fxa_basket_export
The service account is only granted access to create objects,
not to list or delete, so the gcs_delete step was failing.

It seems reasonable to maintain those limited privs and to remove the
delete step. We can revisit later if we find this causes issues in reprocessing.
I believe the overwrite behavior should be sufficient.