Pin pip to <20.3.0 for now until we fix resolution issue (#1198)
Currently, `dataproc_init.sh` updates pip to 20.3.1 which tries to
install the dependencies, but takes > 10 minutes to figure out a
solution set that satisfies the version dependencies.

This has us go back to the previous pip which has the old resolver which
works while we figure out how to straighten out the dependency
requirements.