Execute copy_deduplicate for main_v4 as a single query
We have previously been unable to run copy_deduplicate as a single query
for main_v4 due to resource constraints in BigQuery causing queries to fail.
See https://github.com/mozilla/telemetry-airflow/pull/1224 as an example.

BQ engineering, though, has made some backend changes to improve efficiency
for cases like this, and recent testing has shown that running a single query
is effective. This should be somewhat more cost efficient than splitting the
query into slices, and it can potentially lead to more effective clustering
and less fragmentation.

I will plan to immediately rerun a day once this is merged as a sanity check
that the configuration is correct. If we do see a failure in
coming days or weeks, we can revert to the previous configuration.