Change time of Merino job to Tuesdays (#1663)
* Change time of Merino job to Tuesdays

The job keeps failing on Monday because the upstream Wikimedia export
is not ready when we try to download the file. So changing the date
to Tuesday to improve our chances of download.

* Lint merino_jobs file with black and ruff

* Set Ruff's target Python to 3.8

There have been many changes in type annotations from 3.8
to more modern versions of Python. Some of those best practices
are not valid Python 3.8 syntax. Setting this should ignore
those errors.