feat: Default reattach_on_restart parameter to True in bigquery_etl_query function (#1961)
We've been hit (again) by Airflow losing track of a running pod, restarting a task and failing with `More than one pod running with labels ...` error.
We previously set some of the tasks to reattach to running pods on restart ([1], [2], [3]) which solved this problem for them. We didn't set it by default for all tasks back then to avoid situations where not-autodeleted pods are stuck with the old code. Now that DEs have permissions required to delete pods it seems safe to revisit this.

[1] https://github.com/mozilla/telemetry-airflow/pull/1820
[2] https://github.com/mozilla/telemetry-airflow/pull/1834
[3] https://github.com/mozilla/telemetry-airflow/pull/1837