Bug 1910995 - Enable the detection of TB metrics
As called out in mozilla/probe-scraper#790 this PR is necessary to address the lack of a way to run the GH workflow for Thunderbird ([docs](https://mozilla.github.io/glean/book/user/adding-glean-to-your-project/enable-data-ingestion.html#validate-and-publish-metrics)). Given that TB uses Mercurial/Git Mirror, I'm replicating the same mechanism used for Firefox Desktop.